backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocg-backend"
version = "0.1.0"
description = "Open Context Graph backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.0",
  "pyjwt[crypto]>=2.9.0",
  "prometheus-client>=0.21.0",
  "typer>=0.12.5",
  "redis>=5.2.0",
  "rq>=1.16.2",
  "httpx>=0.27.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.7.3",
  "mypy>=1.13.0"
]

[project.scripts]
ocg = "ocg.cli:app"

[tool.setuptools]
packages = ["ocg"]

[tool.setuptools.package-dir]
ocg = "ocg"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
